requirements.txt
Flask
requests
feedparser
python-telegram-bot
python-dotenv

